🎨 Palette: Navigation UX and CI Stability Fix
💡 What:
- Refactored Navbar for accessibility (Skip link, NavLink, active styles).
- Resolved project-wide ESLint errors blocking CI (ServiceWorker globals, Node globals, Fast Refresh warnings, unused variables, unescaped entities).
- Migrated CI workflows to use pnpm and added missing dependency steps for security scans.
- Extracted useToast hook to a dedicated file for better architecture and Fast Refresh compatibility.

🎯 Why:
- Previous CI runs were blocked by pre-existing lint errors and tool-mismatch (npm vs pnpm).
- Improves navigation accessibility for keyboard and color-blind users.
- Ensures a clean and maintainable codebase.

♿ Accessibility:
- Added Skip to Content link.
- Non-color-only active states for navigation.
- Fixed unescaped entities in ErrorBoundary.
- Improved focus visible states.
