+++ scripts/test-openrouter.mjs
@@ -1,3 +1,4 @@
+/* eslint-env node */
 import fetch from 'node-fetch';
 import dotenv from 'dotenv';
 import path from 'path';
