+++ src/components/ui/Toaster.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react'
 
-interface Toast {
+export interface Toast {
   id: string
   message: string
   type: 'success' | 'error' | 'info'
@@ -8,21 +8,6 @@ interface Toast {
 
 interface ToasterProps {
   duration?: number
-}
-
-export const useToast = () => {
-  const [toasts, setToasts] = useState<Toast[]>([])
-
-  const addToast = (message: string, type: Toast['type'] = 'info') => {
-    const id = Math.random().toString(36).substr(2, 9)
-    setToasts((prev) => [...prev, { id, message, type }])
-  }
-
-  const removeToast = (id: string) => {
-    setToasts((prev) => prev.filter((toast) => toast.id !== id))
-  }
-
-  return { addToast, removeToast }
 }
 
 export const Toaster = ({ duration = 3000 }: ToasterProps) => {
