+++ src/test/AllTheProviders.tsx
@@ -1,0 +1,9 @@
+import { BrowserRouter } from 'react-router-dom';
+
+export const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <BrowserRouter>
+      {children}
+    </BrowserRouter>
+  );
+};
