+++ src/test/test-utils.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
-import { BrowserRouter } from 'react-router-dom';
 import { render, RenderOptions } from '@testing-library/react';
 import { vi } from 'vitest';
+import { AllTheProviders } from './AllTheProviders';
 
 // Mock Supabase client
 vi.mock('@supabase/supabase-js', () => ({
@@ -21,14 +21,6 @@ vi.mock('@supabase/supabase-js', () => ({
     }),
   }),
 }));
-
-const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <BrowserRouter>
-      {children}
-    </BrowserRouter>
-  );
-};
 
 const customRender = (
   ui: ReactElement,
